+++ tests/ui/sync-fail.rs
@@ -16,8 +16,6 @@ ram_storage!(
     block_size=256,
     block_count=512,
     lookaheadwords_size_ty=consts::U2,
-    filename_max_plus_one_ty=consts::U256,
-    path_max_plus_one_ty=consts::U256,
     result=Result,
 );
 
@@ -32,8 +30,6 @@ ram_storage!(
     block_size=20*35,
     block_count=32,
     lookaheadwords_size_ty=consts::U2,
-    filename_max_plus_one_ty=consts::U256,
-    path_max_plus_one_ty=consts::U256,
     result=Result,
 );